# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  requests.ID:
    properties:
      id:
        type: string
    required:
    - id
    type: object
...
  requests.Pagination:
    properties:
      page:
        minimum: 1
        type: integer
    required:
    - page
    type: object
...
  requests.Search:
    properties:
      page:
        minimum: 1
        type: integer
      search:
        type: string
    required:
    - page
    - search
    type: object
...
  requests.SortFilterTVSeries:
    properties:
      firstAirDateFrom:
        maximum: 2050
        minimum: 1900
        type: integer
      firstAirDateTo:
        maximum: 2050
        minimum: 1900
        type: integer
      genres:
        type: string
      numSeason:
        minimum: 1
        type: integer
      page:
        minimum: 1
        type: integer
      productionCompanies:
        type: string
      sort:
        enum:
        - top
        - popularity
        - new
        - old
        type: string
      status:
        enum:
        - production
        - ended
        - airing
        type: string
    required:
    - page
    - sort
    type: object
...
  responses.Actor:
    properties:
      character:
        type: string
      image:
        type: string
      name:
        type: string
      tmdb_id:
        type: string
    type: object
...
  responses.Author:
    properties:
      _id:
        type: string
      email:
        type: string
      image:
        type: string
      is_premium:
        type: boolean
      username:
        type: string
    type: object
  responses.ConsumeLater:
    properties:
      _id:
        type: string
      content:
        $ref: '#/definitions/responses.ConsumeLaterContent'
      content_external_id:
        type: string
      content_external_int_id:
        type: integer
      content_id:
        type: string
      content_type:
        type: string
      created_at:
        type: string
      self_note:
        type: string
      user_id:
        type: string
    type: object
  responses.ConsumeLaterContent:
    properties:
      description:
        type: string
      image_url:
        type: string
      score:
        type: number
      title_en:
        type: string
      title_original:
        type: string
    type: object
...
  responses.DayOfWeekTVSeries:
    properties:
      data:
        items:
          $ref: '#/definitions/responses.TVSeries'
        type: array
      day_of_week:
        type: integer
    type: object
...
  responses.ProductionAndCompany:
    properties:
      logo:
        type: string
      name:
        type: string
      origin_country:
        type: string
    type: object
  responses.Recommendation:
    properties:
      description:
        type: string
      image_url:
        type: string
      release_date:
        type: string
      title_en:
        type: string
      title_original:
        type: string
      tmdb_id:
        type: string
    type: object
  responses.Review:
    properties:
      _id:
        type: string
      author:
        $ref: '#/definitions/responses.Author'
      content_external_id:
        type: string
      content_external_int_id:
        type: integer
      content_id:
        type: string
      created_at:
        type: string
      is_author:
        type: boolean
      is_liked:
        type: boolean
      likes:
        items:
          type: string
        type: array
      popularity:
        type: integer
      review:
        type: string
      star:
        type: integer
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
  responses.ReviewSummary:
    properties:
      avg_star:
        type: number
      is_reviewed:
        type: boolean
      review:
        $ref: '#/definitions/responses.Review'
      star_counts:
        $ref: '#/definitions/responses.StarCounts'
      total_votes:
        type: integer
    type: object
...
  responses.Season:
    properties:
      air_date:
        type: string
      episode_count:
        type: integer
      image_url:
        type: string
      name:
        type: string
      season_num:
        type: integer
    type: object
  responses.StarCounts:
    properties:
      five_star:
        type: integer
      four_star:
        type: integer
      one_star:
        type: integer
      three_star:
        type: integer
      two_star:
        type: integer
    type: object
  responses.Streaming:
    properties:
      buy_options:
        items:
          $ref: '#/definitions/responses.StreamingPlatform'
        type: array
      country_code:
        type: string
      rent_options:
        items:
          $ref: '#/definitions/responses.StreamingPlatform'
        type: array
      streaming_platforms:
        items:
          $ref: '#/definitions/responses.StreamingPlatform'
        type: array
    type: object
  responses.StreamingPlatform:
    properties:
      logo:
        type: string
      name:
        type: string
    type: object
  responses.TVDetailsList:
    properties:
      _id:
        type: string
      created_at:
        type: string
      score:
        type: number
      status:
        type: string
      times_finished:
        type: integer
      tv_id:
        type: string
      tv_tmdb_id:
        type: string
      user_id:
        type: string
      watched_episodes:
        type: integer
      watched_seasons:
        type: integer
    type: object
  responses.TVSeries:
    properties:
      _id:
        type: string
      actors:
        items:
          $ref: '#/definitions/responses.Actor'
        type: array
      backdrop:
        type: string
      description:
        type: string
      first_air_date:
        type: string
      genres:
        items:
          type: string
        type: array
      image_url:
        type: string
      images:
        items:
          type: string
        type: array
      networks:
        items:
          $ref: '#/definitions/responses.ProductionAndCompany'
        type: array
      production_companies:
        items:
          $ref: '#/definitions/responses.ProductionAndCompany'
        type: array
      recommendations:
        items:
          $ref: '#/definitions/responses.Recommendation'
        type: array
      reviews:
        $ref: '#/definitions/responses.ReviewSummary'
      seasons:
        items:
          $ref: '#/definitions/responses.Season'
        type: array
      status:
        type: string
      streaming:
        items:
          $ref: '#/definitions/responses.Streaming'
        type: array
      title_en:
        type: string
      title_original:
        type: string
      tmdb_id:
        type: string
      tmdb_popularity:
        type: number
      tmdb_vote:
        type: number
      tmdb_vote_count:
        type: integer
      top_rated:
        type: number
      total_episodes:
        type: integer
      total_seasons:
        type: integer
      translations:
        items:
          $ref: '#/definitions/responses.Translation'
        type: array
      videos:
        items:
          $ref: '#/definitions/responses.Trailer'
        type: array
    type: object
  responses.TVSeriesDetails:
    properties:
      _id:
        type: string
      actors:
        items:
          $ref: '#/definitions/responses.Actor'
        type: array
      backdrop:
        type: string
      description:
        type: string
      first_air_date:
        type: string
      genres:
        items:
          type: string
        type: array
      image_url:
        type: string
      images:
        items:
          type: string
        type: array
      networks:
        items:
          $ref: '#/definitions/responses.ProductionAndCompany'
        type: array
      production_companies:
        items:
          $ref: '#/definitions/responses.ProductionAndCompany'
        type: array
      recommendations:
        items:
          $ref: '#/definitions/responses.Recommendation'
        type: array
      reviews:
        $ref: '#/definitions/responses.ReviewSummary'
      seasons:
        items:
          $ref: '#/definitions/responses.Season'
        type: array
      status:
        type: string
      streaming:
        items:
          $ref: '#/definitions/responses.Streaming'
        type: array
      title_en:
        type: string
      title_original:
        type: string
      tmdb_id:
        type: string
      tmdb_popularity:
        type: number
      tmdb_vote:
        type: number
      tmdb_vote_count:
        type: integer
      total_episodes:
        type: integer
      total_seasons:
        type: integer
      translations:
        items:
          $ref: '#/definitions/responses.Translation'
        type: array
      tv_list:
        $ref: '#/definitions/responses.TVDetailsList'
      videos:
        items:
          $ref: '#/definitions/responses.Trailer'
        type: array
      watch_later:
        $ref: '#/definitions/responses.ConsumeLater'
    type: object
...
  responses.Trailer:
    properties:
      key:
        type: string
      name:
        type: string
      title_original:
        type: string
    type: object
  responses.Translation:
    properties:
      description:
        type: string
      lan_code:
        type: string
      lan_name:
        type: string
      lan_name_en:
        type: string
      title:
        type: string
    type: object
...
host: http://localhost:8080
info:
  contact:
    email: mrntlu@gmail.com
    name: Burak Fidan
  description: REST Api of Project Consumer.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: '#soon'
  title: Project Consumer API
  version: "1.0"
paths:
...
  /tv:
    get:
      consumes:
      - application/json
      description: Returns tv series by sort and filter with pagination
      parameters:
      - description: Sort and Filter TV Series
        in: body
        name: sortfiltertvseries
        required: true
        schema:
          $ref: '#/definitions/requests.SortFilterTVSeries'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/responses.TVSeries'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get TV Series by Sort and Filter
      tags:
      - tv
  /tv/airing:
    get:
      consumes:
      - application/json
      description: Returns list of tv series by day of week
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/responses.DayOfWeekTVSeries'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get Currently Airing TV Series by day of week and country code
      tags:
      - tv
  /tv/details:
    get:
      consumes:
      - application/json
      description: Returns tv series details with optional authentication
      parameters:
      - description: ID
        in: body
        name: id
        required: true
        schema:
          $ref: '#/definitions/requests.ID'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/responses.TVSeriesDetails'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get TV Series Details
      tags:
      - tv
  /tv/search:
    get:
      consumes:
      - application/json
      description: Search tv series
      parameters:
      - description: Search
        in: body
        name: search
        required: true
        schema:
          $ref: '#/definitions/requests.Search'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/responses.TVSeries'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Search TV Series
      tags:
      - tv
  /tv/upcoming:
    get:
      consumes:
      - application/json
      description: Returns upcoming tv series by sort with pagination
      parameters:
      - description: Pagination
        in: body
        name: pagination
        required: true
        schema:
          $ref: '#/definitions/requests.Pagination'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/responses.TVSeries'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get Upcoming TV Series by Sort
      tags:
      - tv
...
schemes:
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"